[QuestNames] added support for showing questname upon starting a conversation
Questnames can be configured in a conversation with the "questname" value at the top level, or in a starting npc-option.

Case1: an npc is used for one quest
Solution: add questname: 'My Quest Name' below the "quester" value (top-level)

Case2: an npc is used for a quest but also has a dialogue fragment post-quest
Solution: add the questname value to each starting npc_option that belongs to the quest and leave the ones post-quest without one, and also leave out the top-level questname

Case3: an npc is used for multiple quests
Solution: add the appropriate questname value to each starting npc_option and leave the top-level questname empty or out

Extra note: if the questname is specified at the conversation top-level as well as in a starting npc_option, the npc_option one will override the top-level one. However this way you cannot override with the empty string, for that case3 should be used.