[QuestNames] fixed backwards compatibility prior to questnames
When a conversation doesn't specify a questname or the empty string, it will send the old conversation_start message without a name argument just like before.

If the conversation or the starting npc_option or both does specify a questname, it will now show a new conversation_start_named message, that takes two arguments, adding the questname as {2}. This new message has been made available to all language config files.