Make all tags and points package dependent
This is achieved by adding package prefix to all tags and points in all
packages with "tag_point_prefix" set to false. Tags and points that are
used in more than one such package are considered global and they will
be prefixed with "global" package name. If "global" package already
exists and is used, the plugin will find another name, like "global-4"
to avoid conflicts.

Tags and points are updated in the events.yml, conditions.yml, main.yml
(quest cancelers) files and in the database. Tag and point
events/conditions are considered inside and outside of "check"
condition/"run" event.