Created a MythicHider file that deals with the hiding of mythicMobs in a similar way to the NPCHider. Limitations currently are that mob sounds or particles are not hidden. To allow for the usage of the hider added three arguments to the MythicSpawnMobEvent file. 1: private, this will hide the mob for all but the player who the command is run for. 2: target, this will cause the mob to target the player when it spawns. Further targetting issues should be dealt with in the configuration of the mob. 3: marked, similar to the regular mob spawn mobs spawned can now be marked and thus we can track if the private mob is killed by a player. Thus facilitating the next change.
Added checking marked for the MythicMobKillObjective.

Lastly added a check in the MythicMobsIntegrator for if protocollib is hooked, if it is we then start the MythicHider process.

Further Improvements could be made to the hider as it currently is not persistent and will reset on every restart. Ideally in the future one could specify Mythic Mobs as we do for NPCs now however my current implementation uses no form of storage and thus will be reset.
Added changes to changelog

Created a MythicHider file that deals with the hiding of mythicMobs in a similar way to the NPCHider. Limitations currently are that mob sounds or particles are not hidden.
To allow for the usage of the hider added three arguments to the MythicSpawnMobEvent file.
1: private, this will hide the mob for all but the player who the command is run for.
2: target, this will cause the mob to target the player when it spawns. Further targetting issues should be dealt with in the configuration of the mob.
3: marked, similar to the regular mob spawn mobs spawned can now be marked and thus we can track if the private mob is killed by a player. Thus facilitating the next change.

Added checking marked for the MythicMobKillObjective.

Lastly added a check in the MythicMobsIntegrator for if protocollib is hooked, if it is we then start the MythicHider process.

Further Improvements could be made to the hider as it currently is not persistent and will reset on every restart. Ideally in the future one could specify Mythic Mobs as we do for NPCs now however my current implementation uses no form of storage and thus will be reset.

Created a MythicHider file that deals with the hiding of mythicMobs in a similar way to the NPCHider. Limitations currently are that mob sounds or particles are not hidden.
To allow for the usage of the hider added three arguments to the MythicSpawnMobEvent file.
1: private, this will hide the mob for all but the player who the command is run for.
2: target, this will cause the mob to target the player when it spawns. Further targetting issues should be dealt with in the configuration of the mob.
3: marked, similar to the regular mob spawn mobs spawned can now be marked and thus we can track if the private mob is killed by a player. Thus facilitating the next change.

Added checking marked for the MythicMobKillObjective.

Lastly added a check in the MythicMobsIntegrator for if protocollib is hooked, if it is we then start the MythicHider process.

Further Improvements could be made to the hider as it currently is not persistent and will reset on every restart. Ideally in the future one could specify Mythic Mobs as we do for NPCs now however my current implementation uses no form of storage and thus will be reset.

Created a MythicHider file that deals with the hiding of mythicMobs in a similar way to the NPCHider. Limitations currently are that mob sounds or particles are not hidden.
To allow for the usage of the hider added three arguments to the MythicSpawnMobEvent file.
1: private, this will hide the mob for all but the player who the command is run for.
2: target, this will cause the mob to target the player when it spawns. Further targetting issues should be dealt with in the configuration of the mob.
3: marked, similar to the regular mob spawn mobs spawned can now be marked and thus we can track if the private mob is killed by a player. Thus facilitating the next change.

Added checking marked for the MythicMobKillObjective.

Lastly added a check in the MythicMobsIntegrator for if protocollib is hooked, if it is we then start the MythicHider process.

Further Improvements could be made to the hider as it currently is not persistent and will reset on every restart. Ideally in the future one could specify Mythic Mobs as we do for NPCs now however my current implementation uses no form of storage and thus will be reset.
Added changes to changelog

Created a MythicHider file that deals with the hiding of mythicMobs in a similar way to the NPCHider. Limitations currently are that mob sounds or particles are not hidden.
To allow for the usage of the hider added three arguments to the MythicSpawnMobEvent file.
1: private, this will hide the mob for all but the player who the command is run for.
2: target, this will cause the mob to target the player when it spawns. Further targetting issues should be dealt with in the configuration of the mob.
3: marked, similar to the regular mob spawn mobs spawned can now be marked and thus we can track if the private mob is killed by a player. Thus facilitating the next change.

Added checking marked for the MythicMobKillObjective.

Lastly added a check in the MythicMobsIntegrator for if protocollib is hooked, if it is we then start the MythicHider process.

Further Improvements could be made to the hider as it currently is not persistent and will reset on every restart. Ideally in the future one could specify Mythic Mobs as we do for NPCs now however my current implementation uses no form of storage and thus will be reset.
Added changes to changelog
Added changes to documentation

Created a MythicHider file that deals with the hiding of mythicMobs in a similar way to the NPCHider. Limitations currently are that mob sounds or particles are not hidden.
To allow for the usage of the hider added three arguments to the MythicSpawnMobEvent file.
1: private, this will hide the mob for all but the player who the command is run for.
2: target, this will cause the mob to target the player when it spawns. Further targetting issues should be dealt with in the configuration of the mob.
3: marked, similar to the regular mob spawn mobs spawned can now be marked and thus we can track if the private mob is killed by a player. Thus facilitating the next change.

Added checking marked for the MythicMobKillObjective.

Lastly added a check in the MythicMobsIntegrator for if protocollib is hooked, if it is we then start the MythicHider process.

Further Improvements could be made to the hider as it currently is not persistent and will reset on every restart. Ideally in the future one could specify Mythic Mobs as we do for NPCs now however my current implementation uses no form of storage and thus will be reset.
Added changes to changelog