Re-introduce raw instruction field in Instruction class.
At the moment it is not possible to replace the getInstruction() method
with an implementation that generates the instruction string from the
tokenized instruction part. This is because quite a lot of classes use
the raw instruction instead of letting the instruction class to the
tokenizing and parsing. Examples are Conversations, the run event and
the notify event.